pyproject.toml
[project]
name = "beam-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
